requirements.txt
django
PyJWT
djangorestframework-jwt
